Rename SykmeldingSide to SykmeldingPage and move sykmeldinger components into separate folder
