+++ client/src/App.tsx
@@ -3,7 +3,7 @@ import './App.less';
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
-import SykmeldingSide from './pages/sykmelding/SykmeldingSide';
+import SykmeldingSide from './pages/sykmelding/SykmeldingPage';
 import SykmeldingerPage from './pages/sykmeldinger/SykmeldingerPage';
 
 const App = () => {
