+++ client/src/pages/sykmeldinger/components/Lenkepanel/Lenkepanel.less
