+++ client/src/pages/sykmeldinger/components/LenkepanelContainer.less
